+++ griffon/commands/queries.py
@@ -275,6 +275,13 @@ def retrieve_component_summary(ctx, component_name, strict_name_search):
     help="Include components from inactive product streams.",
 )
 @click.option(
+    "--include-product-streams-excluded-components",
+    "include_product_stream_excluded_components",
+    is_flag=True,
+    default=get_config_option("default", "include_product_stream_excluded_components", False),
+    help="Include product stream excluded components.",
+)
+@click.option(
     "-v",
     "verbose",
     count=True,
@@ -304,6 +311,7 @@ def get_product_contain_component(
     no_community,
     no_middleware,
     include_inactive_product_streams,
+    include_product_stream_excluded_components,
     verbose,
 ):
     with console.status("griffoning", spinner="line") as operation_status:
@@ -407,7 +415,11 @@ def get_product_contain_component(
                     ctx.obj["PROFILE"], "exclude_components"
                 ).split("\n")
             normalised_results = generate_normalised_results(
-                output, exclude_products, exclude_components, include_inactive_product_streams
+                output,
+                exclude_products,
+                exclude_components,
+                include_inactive_product_streams,
+                include_product_stream_excluded_components,
             )
             result_tree = generate_result_tree(normalised_results)
             affects = generate_affects(ctx, result_tree, exclude_components, "add", format="json")